**** Revision 358:
various casts to fix compiler warnings
fixed a memory leak in msxproj.c
Forced LoadLibrary and CreateProcess function calls to use non-unicode versions by calling LoadLibraryA and CreateProcessA.
Modified logic in MSXcompiler_open function to close the compiler even if the compile failed.
added *.user to svn:ignore for /build/MSVC/epanet-msx-dll directory to prevent msvc user settings from accidentally getting committed to the repository.
Added check to make sure that DLLEXPORT isn't already defined in epanetmsx.h

+++ src/mathexpr.c
@@ -733,7 +733,7 @@ MathExpr * mathexpr_create(char *formula, int (*getVar) (char *))
     PrevLex = 0;
     CurLex = 0;
     S = formula;
-    Len = strlen(S);
+    Len = (int)strlen(S);
     Pos = 0;
     Bc = 0;
     tree = getTree();
