+++ src/msxout.c
@@ -123,7 +123,7 @@ int MSXout_saveInitialResults()
     fwrite(&n, sizeof(INT4), 1, f);                         //Reporting step size
     for (m=1; m<=MSX.Nobjects[SPECIES]; m++)
     {
-        n = strlen(MSX.Species[m].id);
+        n = (int)strlen(MSX.Species[m].id);
         fwrite(&n, sizeof(INT4), 1, f);                     //Length of species ID
         fwrite(MSX.Species[m].id, sizeof(char), n, f);      //Species ID string
     }
@@ -158,7 +158,6 @@ int MSXout_saveResults()
 {
     int   m, j;
     REAL4 x;
-
     for (m=1; m<=MSX.Nobjects[SPECIES]; m++)
     {
         for (j=1; j<=MSX.Nobjects[NODE]; j++)
